+++ nyquist/nylsf/alaw.c
@@ -366,7 +366,7 @@ alaw_read_alaw2s (SF_PRIVATE *psf, short *ptr, sf_count_t len)
 	while (len > 0)
 	{	if (len < bufferlen)
 			bufferlen = (int) len ;
-		readcount = psf_fread (psf->u.ucbuf, 1, bufferlen, psf) ;
+		readcount = (int) psf_fread (psf->u.ucbuf, 1, bufferlen, psf) ;
 		alaw2s_array (psf->u.ucbuf, readcount, ptr + total) ;
 		total += readcount ;
 		if (readcount < bufferlen)
@@ -387,7 +387,7 @@ alaw_read_alaw2i (SF_PRIVATE *psf, int *ptr, sf_count_t len)
 	while (len > 0)
 	{	if (len < bufferlen)
 			bufferlen = (int) len ;
-		readcount = psf_fread (psf->u.ucbuf, 1, bufferlen, psf) ;
+		readcount = (int) psf_fread (psf->u.ucbuf, 1, bufferlen, psf) ;
 		alaw2i_array (psf->u.ucbuf, readcount, ptr + total) ;
 		total += readcount ;
 		if (readcount < bufferlen)
@@ -411,7 +411,7 @@ alaw_read_alaw2f (SF_PRIVATE *psf, float *ptr, sf_count_t len)
 	while (len > 0)
 	{	if (len < bufferlen)
 			bufferlen = (int) len ;
-		readcount = psf_fread (psf->u.ucbuf, 1, bufferlen, psf) ;
+		readcount = (int) psf_fread (psf->u.ucbuf, 1, bufferlen, psf) ;
 		alaw2f_array (psf->u.ucbuf, readcount, ptr + total, normfact) ;
 		total += readcount ;
 		if (readcount < bufferlen)
@@ -434,7 +434,7 @@ alaw_read_alaw2d (SF_PRIVATE *psf, double *ptr, sf_count_t len)
 	while (len > 0)
 	{	if (len < bufferlen)
 			bufferlen = (int) len ;
-		readcount = psf_fread (psf->u.ucbuf, 1, bufferlen, psf) ;
+		readcount = (int) psf_fread (psf->u.ucbuf, 1, bufferlen, psf) ;
 		alaw2d_array (psf->u.ucbuf, readcount, ptr + total, normfact) ;
 		total += readcount ;
 		if (readcount < bufferlen)
@@ -459,7 +459,7 @@ alaw_write_s2alaw	(SF_PRIVATE *psf, const short *ptr, sf_count_t len)
 	{	if (len < bufferlen)
 			bufferlen = (int) len ;
 		s2alaw_array (ptr + total, bufferlen, psf->u.ucbuf) ;
-		writecount = psf_fwrite (psf->u.ucbuf, 1, bufferlen, psf) ;
+		writecount = (int) psf_fwrite (psf->u.ucbuf, 1, bufferlen, psf) ;
 		total += writecount ;
 		if (writecount < bufferlen)
 			break ;
@@ -480,7 +480,7 @@ alaw_write_i2alaw	(SF_PRIVATE *psf, const int *ptr, sf_count_t len)
 	{	if (len < bufferlen)
 			bufferlen = (int) len ;
 		i2alaw_array (ptr + total, bufferlen, psf->u.ucbuf) ;
-		writecount = psf_fwrite (psf->u.ucbuf, 1, bufferlen, psf) ;
+		writecount = (int) psf_fwrite (psf->u.ucbuf, 1, bufferlen, psf) ;
 		total += writecount ;
 		if (writecount < bufferlen)
 			break ;
@@ -504,7 +504,7 @@ alaw_write_f2alaw	(SF_PRIVATE *psf, const float *ptr, sf_count_t len)
 	{	if (len < bufferlen)
 			bufferlen = (int) len ;
 		f2alaw_array (ptr + total, bufferlen, psf->u.ucbuf, normfact) ;
-		writecount = psf_fwrite (psf->u.ucbuf, 1, bufferlen, psf) ;
+		writecount = (int) psf_fwrite (psf->u.ucbuf, 1, bufferlen, psf) ;
 		total += writecount ;
 		if (writecount < bufferlen)
 			break ;
@@ -528,7 +528,7 @@ alaw_write_d2alaw	(SF_PRIVATE *psf, const double *ptr, sf_count_t len)
 	{	if (len < bufferlen)
 			bufferlen = (int) len ;
 		d2alaw_array (ptr + total, bufferlen, psf->u.ucbuf, normfact) ;
-		writecount = psf_fwrite (psf->u.ucbuf, 1, bufferlen, psf) ;
+		writecount = (int) psf_fwrite (psf->u.ucbuf, 1, bufferlen, psf) ;
 		total += writecount ;
 		if (writecount < bufferlen)
 			break ;
